add google api geocoding functionality

+++ src/main/java/com/hcl/retailshop/model/Coordinates.java
@@ -1,6 +1,8 @@
 package com.hcl.retailshop.model;
 
 /**
+ * The Coordinates object to store longitude and latitude.
+ *
  * Created by Nitin Shreyakar.
  */
 public class Coordinates {
@@ -8,23 +10,23 @@ public class Coordinates {
     /**
      * The geocode longitude.
      */
-    private final Integer longitude;
+    private final double longitude;
 
     /**
      * The geocode latitude.
      */
-    private final Integer latitude;
+    private final double latitude;
 
-    public Coordinates(Integer longitude, Integer latitude) {
+    public Coordinates(double longitude, double latitude) {
         this.longitude = longitude;
         this.latitude = latitude;
     }
 
-    public Integer getLongitude() {
+    public double getLongitude() {
         return longitude;
     }
 
-    public Integer getLatitude() {
+    public double getLatitude() {
         return latitude;
     }
 }
